auto preview clean up
